Move jtorch override into wan_tx
Don't need to override xla/torchax package anymore
